Make logical switch ports numbers and names independent - linc part
This commit separates capable switch ports' numbers from logical switch
ports' numbers. It also introduces a possibility to set custom name for
logical switch ports. The above is achieved by new options for logical
switch ports: `port_no` and `port_name`. They should be used as follows:
{port, 2, [{queues, []}, %% 2 is the capable switch port number
          {port_no, 1},  %% 1 is the logical port number
          {port_name, "Banshee"}]}

NOTE: Now the third element of the port tuple is a list but the old
format with {queues, QueuesConfig} as the third element is still
acceptable.

If the options are not provided `port_no` defaults to capable port number
and `port_name` to `Port{CAPABLE_SWITCH_PORT_NO}` or
`Port{LOGICAL_SWITCH_PORT_NO}` if `port_no` option is not set.

Also some refactorings are introduced and description for new options
in the sys.config.orig.